+++ frontend/src/pages/authentication/index.ts
@@ -1,0 +1,4 @@
+import Signin from './signin';
+import Signup from './signup';
+
+export { Signin, Signup };
