+++ frontend/src/pages/dashboard/index.ts
@@ -1,0 +1,3 @@
+import Dashboard from './Dashboard';
+
+export default Dashboard;
