--- frontend/src/pages/home/Home.tsx
@@ -1,5 +1,0 @@
-const Home = () => {
-  return <h1>Money Super Hero</h1>;
-};
-
-export default Home;
