--- frontend/src/pages/home/index.ts
@@ -1,4 +1,0 @@
-import Home from './Home';
-
-
-export default Home;
