+++ frontend/src/pages/settings/index.ts
@@ -1,4 +1,3 @@
 import Settings from './Settings';
 
-
-export { Settings };
+export default Settings;
